Very neat addition of triangles
Adding in a similar thing to the bass but for the midrange now. Not sure if it will stay like this and i'd image the current state could induce a few seizures but it looks cool. Basically changed bass to spawn in less frequently and added in a triangle spiral goping down the middle responding to the midrange and spawning as you go. Havent made a new material for it yet but looks cool so far. I had to make a hollow triangle prefab and for that i just used probuilders experimental boolean. worked well.