fixed some emission problems
Previously, the emission colour of the object was being set by the audiobuffer value. this was always white to black.  I have now made it so it finds the emission colour set on the prefab, subtracts each component from 1 and then uses that difference to subtract from the audiobuffer value so it retains its original emission colur, just changes the intensity of it.