Making tunnel longer and shorter with bass
So ive made the tunnels get shorter and longer with the beat. Some problems occured due to the spiral nature of the triangle tunnel, in that when I remake the tunnel the spiral is sometimes slightly offset. Figured out a semi-workaround but not a proper fix. I'm not 100% sure whats causing the problem but i think its something to do with the destroying of the originals in that its not happening fast enough. Idk quite tired so last commit of night.