lots of cool stuff
I added in the ball in the centre on the tunnel with a trail. Its supposed to cool like the cover of the album "Currents" and I think it actually looks really nice. I'm gonna upload a picture of the album cover and a screenshot of the application now along with this commit.